Fixed writing to SD card.

+++ 2016-Competition-AirBrakes/src/CppVersion/Loop/SDCard.h
@@ -26,7 +26,7 @@ public:
 	// -----------------------------------------------------------------------------------
 	// This methods writes data to the SD card data file.
 	// -----------------------------------------------------------------------------------
-	void WriteToSD(int deltaTime, float dis, float vel, float ax, float ay, float az, float filteredDis, float filteredVel, float filteredAx, int airBrakePercent);
+	void WriteToSD(float deltaTime, float dis, float vel, float ax, float ay, float az, float filteredDis, float filteredVel, float filteredAx, int airBrakePercent);
 	// -----------------------------------------------------------------------------------
 	// This methods reads the apogee value from the SD card. 
 	// It looks for a file name apogee.txt on the SD card. If it finds one, it reads
